add migration and admin page

+++ src/main/java/ru/wolves/bookingsite/repositories/RoomHallRepo.java
@@ -7,7 +7,7 @@ import ru.wolves.bookingsite.models.RoomHall;
 import java.util.List;
 import java.util.Optional;
 
-public interface RoomHallRepo extends JpaRepository<RoomHall,Integer> {
+public interface RoomHallRepo extends JpaRepository<RoomHall,Long> {
     Optional<RoomHall> findByName(String name);
     Optional<RoomHall> findByEngName(String engName);
 }
