+++ src/main/java/ru/wolves/bookingsite/services/PersonService.java
@@ -5,9 +5,9 @@ import ru.wolves.bookingsite.models.Person;
 import java.util.List;
 
 public interface PersonService {
-    Person findPerson(int id);
+    Person findPerson(Long id);
     List<Person> findAll();
     Person updatePerson(Person person);
     void savePerson(Person person);
-    void deletePerson(int id);
+    void deletePerson(Long id);
 }
