+++ src/main/java/ru/wolves/bookingsite/services/RoomHallService.java
@@ -6,14 +6,14 @@ import ru.wolves.bookingsite.models.RoomHall;
 import java.util.List;
 
 public interface RoomHallService {
-    RoomHall findRoom(int id);
+    RoomHall findRoom(Long id);
     RoomHall findRoom(String name);
     List<RoomHall> findAllRoomHall();
     void saveRoom(RoomHall roomHall);
 
     RoomHall updateRoom(RoomHall roomhall);
 
-    void deleteRoomHall(int id);
+    void deleteRoomHall(Long id);
     List<RoomHall> findFreeRooms(Booking booking);
     RoomHall findByEngName(String engName);
